jl_cockpit

Documentation for jl_cockpit.

Base.push!

Base.push!

jl_cockpit.accuracy

jl_cockpit.loss_plot!

jl_cockpit.preprocess

jl_cockpit.training_loop

jl_cockpit.visualiser

Base.push! — Method

This function updates Observable with datapoint

Base.push! — Method

This function updates list of Observables with specific value of FLoat32

jl_cockpit.accuracy — Method

This function calculates the accuracy of the model on test data

jl_cockpit.loss_plot! — Method

This function plots the loss of the plot data and axis

jl_cockpit.preprocess — Method

This function preprocesses features x and labels y from the given dataset for the training.
It transforms x into a 28x28 matrix and y into a 10-class vector.

jl_cockpit.training_loop — Method

This function trains the data in a loop
Call this function with optional parameters

jl_cockpit.visualiser — Method

visualiser(; vis_loss::Observable{Vector{Float32}} = nothing, vis_grad_norm::Observable{Vector{Float32}} = nothing)

Initialises the visualiser. It will take the given Observables and display a plot that updates live.