Base.append!

Base.push!

Base.push!

JL_Cockpit.iterate_plot_pos

JL_Cockpit.preprocess

JL_Cockpit.preprocess — Method

preprocess(dataset)

Transform x into a 28x28 matrix and y into a 10-class vector.

JL_Cockpit.iterate_plot_pos — Method

iterate_plot_pos(a::Int64, b::Int64)

Iterate the positions in a Figure, so that there are no gaps and the number of rows is never more than the number of columns.

Base.push! — Method

push!(list_obs::Observable{Vector{T}}, value::T) where {T<:Real}

Push number onto a Vector packaged in the given Observable and trigger Observable.

Base.push! — Method

push!(obs::Observable{Vector{Datapoint}}, dp::Datapoint)

Push Datapoint onto a Vector packaged in the given Observable and trigger Observable.

Base.append! — Method

append!(obs::Observable{Vector{T}}, val_vector::Vector{T}) where {T<:Real}

Append a Vector of Real numbers to an Observable and trigger the Observable.