jl_cockpit.loss_plot! — Method

loss_plot!(fig::Makie.Figure, datapoints::Observable{Vector{Datapoint}})

Plot the loss of each Datapoint as a line graph.

jl_cockpit.grad_norm_plot! — Method

grad_norm_plot!(fig::Makie.Figure, datapoints::Observable{Vector{Datapoint}})

Plot the norm of the gradients in each Datapoint as a line graph.

jl_cockpit.hist_1d_plot! — Method

hist_1d_plot!(fig::Makie.Figure, datapoints::Observable{Vector{Datapoint}})

Plot a histogram of gradients in the last Datapoint.

jl_cockpit.params_plot! — Method

params_plot!(fig::Makie.Figure, datapoints::Observable{Vector{Datapoint}})

Plot a histogram of the parameters given in the last Datapoint.

jl_cockpit.distance_plot! — Method

distance_plot!(fig::Makie.Figure, datapoints::Observable{Vector{Datapoint}})

Plot the l2 distance between the parameters in the first and last Datapoint as a point graph.

jl_cockpit.update_size_plot! — Method

update_size_plot!(fig::Makie.Figure, datapoints::Observable{Vector{Datapoint}})

Plot the l2 distance between parameters in the second to last and last Datapoint given.